refactor: Made the handlers a single class
